requirements.txt
django
djangorestframework
dj-database-url
psycopg2-binary
gunicorn
asgiref
sqlparse
pytz
